+++ csatar-plugins/csatar/lang/en/lang.php
@@ -6,6 +6,10 @@
         'name' => 'CSATÁR',
         'description' => 'Plugin for the RMCSSZ\'s CSATÁR project',
         'author' => 'CSATÁR team',
+        'permissions' => [
+            'manageData' => 'Manage Data',
+            'admin' => 'Csatár admin'
+        ],
         'admin' => [
             'general' => [
                 'name' => 'Name',
@@ -208,6 +212,7 @@
                         'testData' => 'Test data',
                         'importData' => 'Import scouts from ECSET',
                         'synchronizePermissionsMatrix' => 'Synchronize permissions matrix',
+                        'location' => 'Location Data'
                     ],
                 ],
                 'seederData' => [
@@ -515,6 +520,7 @@
                 'validationExceptions' => [
                     'dateInTheFuture' => 'The selected Date is in the future.',
                     'submissionDateAfterApprovalDate' => 'The Submission date cannot be after the approval date.',
+                    'noPermissionToApprove' => 'You do not have permission to approve the team report.',
                 ],
                 'submit_start_date' => 'Team report submit period start date',
                 'submit_end_date' => 'Team report submit period end date',
@@ -568,6 +574,10 @@
                 'dataEntry' => 'Accident log data entry group',
                 'admin' => 'Accident log admin group',
             ],
+            'backendUserRoles' => [
+                'backendUserRoles' => 'Backend User Roles',
+                'rmcsszOffice' => 'RMCSSZ office',
+            ],
             'membershipCard' => [
                 'membershipCard' => 'Membership Card',
                 'membershipCards' => 'Membership Cards',
@@ -575,6 +585,7 @@
                 'active' => 'Active',
             ],
         ],
+
         'locations' => [
             'country' => 'Country',
             'county' => 'County',