+++ csatar-plugins/forms/lang/hu/lang.php
@@ -3,6 +3,9 @@
         'name' => 'Űrlapkészítő',
         'description' => 'Űrlapkészítő a felhasználói felülethez',
         'author' => 'Csatár csapat',
+        'permissions' => [
+            'manageData' => 'Űrlapok kezelése',
+        ],
         'admin' => [
             'form' => [
                 'title' => 'Cím',
@@ -47,6 +50,9 @@
                 ]
             ],
             'select' => 'Kiválasztás',
+            '2FAtoRead' => 'A következő információk megjelenítéséhez két faktoros hitelesítés szükséges: ',
+            '2FAtoCreate' => 'A következő információk megadásához két faktoros hitelesítés szükséges: ',
+            '2FAtoModify' => 'A következő információk szerkesztéséhez két faktoros hitelesítés szükséges: ',
         ],
     ],
     'widgets' => [